# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bank Project API Documentation",
        "description": "API endpoints for managing departments, products, and product types in the Bank Project.",
        "contact": {
            "email": "tvoj.email@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Bank Project API Server"
        },
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get all products",
                "description": "Returns a list of all products, including their product type.",
                "operationId": "7cd9d394f726672223f216b77acab086",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Neautoriziran pristup - potrebna Basic HTTP autentifikacija."
                    }
                },
                "security": {
                    "basicAuth": {}
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create a new product",
                "description": "Creates a new product record.",
                "operationId": "e83af3dd12f9ebc4ad755304893c016f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "PRODUCT_CD",
                                    "NAME"
                                ],
                                "properties": {
                                    "PRODUCT_CD": {
                                        "type": "string",
                                        "example": "PROD001"
                                    },
                                    "NAME": {
                                        "type": "string",
                                        "example": "Savings Account"
                                    },
                                    "DATE_OFFERED": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-01-01",
                                        "nullable": true
                                    },
                                    "DATE_RETIRED": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-31",
                                        "nullable": true
                                    },
                                    "PRODUCT_TYPE_CD": {
                                        "description": "Must exist in Product Types",
                                        "type": "string",
                                        "example": "ACC",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Proizvod uspjesno kreiran",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Greska validacije",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation Error"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "PRODUCT_CD": [
                                                    "The PRODUCT_CD has already been taken."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Neautoriziran pristup"
                    }
                },
                "security": {
                    "basicAuth": {}
                }
            }
        },
        "/api/products/{product_cd}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get product by code",
                "description": "Returns a single product by its code, including its product type.",
                "operationId": "ba635fa706071d6fab5492cb113cf7aa",
                "parameters": [
                    {
                        "name": "product_cd",
                        "in": "path",
                        "description": "Code (PRODUCT_CD) of the product to retrieve",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Product not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Neautoriziran pristup"
                    }
                },
                "security": {
                    "basicAuth": {}
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update an existing product",
                "description": "Updates an existing product record by code.",
                "operationId": "2549e1e159785b156648f41ee2069bd9",
                "parameters": [
                    {
                        "name": "product_cd",
                        "in": "path",
                        "description": "Code (PRODUCT_CD) of the product to update",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "NAME"
                                ],
                                "properties": {
                                    "PRODUCT_CD": {
                                        "description": "Can be optionally updated",
                                        "type": "string",
                                        "example": "PROD001"
                                    },
                                    "NAME": {
                                        "type": "string",
                                        "example": "New Savings Account Name"
                                    },
                                    "DATE_OFFERED": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-01-01",
                                        "nullable": true
                                    },
                                    "DATE_RETIRED": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-06-30",
                                        "nullable": true
                                    },
                                    "PRODUCT_TYPE_CD": {
                                        "description": "Must exist in Product Types",
                                        "type": "string",
                                        "example": "CHK",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Product not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation Error"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "NAME": [
                                                    "The NAME field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Neautoriziran pristup"
                    }
                },
                "security": {
                    "basicAuth": {}
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete a product",
                "description": "Deletes a product record by code.",
                "operationId": "d24cca4dddaafe6040e7d5699c12d878",
                "parameters": [
                    {
                        "name": "product_cd",
                        "in": "path",
                        "description": "Code (PRODUCT_CD) of the product to delete",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Product deleted successfully (No Content)"
                    },
                    "404": {
                        "description": "Product not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Product not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Neautoriziran pristup"
                    }
                },
                "security": {
                    "basicAuth": {}
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Product": {
                "title": "Product",
                "description": "Product model",
                "properties": {
                    "PRODUCT_CD": {
                        "title": "Product Code",
                        "description": "Unique code for the product",
                        "type": "string",
                        "example": "PROD001"
                    },
                    "NAME": {
                        "title": "Name",
                        "description": "Name of the product",
                        "type": "string",
                        "example": "Current Account"
                    },
                    "DATE_OFFERED": {
                        "title": "Date Offered",
                        "description": "Date when the product was offered",
                        "type": "string",
                        "format": "date",
                        "example": "2023-01-01",
                        "nullable": true
                    },
                    "DATE_RETIRED": {
                        "title": "Date Retired",
                        "description": "Date when the product was retired",
                        "type": "string",
                        "format": "date",
                        "example": "2025-12-31",
                        "nullable": true
                    },
                    "PRODUCT_TYPE_CD": {
                        "title": "Product Type Code (Foreign Key)",
                        "description": "Foreign key to ProductType",
                        "type": "string",
                        "example": "CHK"
                    },
                    "product_type": {
                        "$ref": "#/components/schemas/ProductType"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Product"
                }
            },
            "ProductType": {
                "title": "ProductType",
                "description": "Product Type model",
                "properties": {
                    "PRODUCT_TYPE_CD": {
                        "title": "Product Type Code",
                        "description": "Unique code for the product type",
                        "type": "string",
                        "example": "ACC"
                    },
                    "NAME": {
                        "title": "Name",
                        "description": "Name of the product type",
                        "type": "string",
                        "example": "Account"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ProductType"
                }
            }
        },
        "securitySchemes": {
            "basicAuth": {
                "type": "http",
                "scheme": "basic",
                "description": "HTTP Basic Authentication (Koristi e-mail i lozinku testnog korisnika)"
            }
        }
    },
...
    "security": [
        {
            "basicAuth": []
        }
    ]
}